Remove reset args as not in use; Correctly use variables
